fixed undefined label
The documentation of subfiles (section 4.2) is straightforward: add the usual label, then prefix it with M- in another subfile when using pageref (this doesn't hurt anything in the same subfile, AFAICT)

compile main twice, then compile subfiles. Now that the label is defined, this works for psalm 112_4E, printed once but needed three times in about 20 pages.